+++ lib/compress.js
@@ -1,4 +1,4 @@
-var SourceMapGenerator = require('source-map').SourceMapGenerator;
+var CompilerBase = require('./compiler');
 
 /**
  * Expose compiler.
@@ -11,55 +11,10 @@ module.exports = Compiler;
  */
 
 function Compiler(options) {
-  options = options || {};
-  this.map = new SourceMapGenerator({file: options.filename || 'generated.css'});
-  this.position = {line: 1, column: 1};
+  CompilerBase.call(this, options);
 }
 
-Compiler.prototype.updatePosition = function(str) {
-  var lines = str.match(/\n/g);
-  if (lines) this.position.line += lines.length;
-  var i = str.lastIndexOf('\n');
-  this.position.column = ~i ? str.length - i : this.position.column + str.length;
-}
-
-/**
- * Emit string and update current position
- */
-
-Compiler.prototype.emit = function(str, pos, startOnly) {
-  if (pos && pos.start) {
-    this.map.addMapping({
-      generated: {
-        line: this.position.line,
-        column: Math.max(this.position.column - 1, 0)
-      },
-      source: 'source.css',
-      original: {
-        line: pos.start.line,
-        column: pos.start.column - 1
-      }
-    });
-  }
-
-  this.updatePosition(str);
-
-  if (!startOnly && pos && pos.end) {
-    this.map.addMapping({
-      generated: {
-        line: this.position.line,
-        column: Math.max(this.position.column - 1, 0)
-      },
-      source: 'source.css',
-      original: {
-        line: pos.end.line,
-        column: pos.end.column - 1
-      }
-    });
-  }
-
-  return str;
-}
+Compiler.prototype = new CompilerBase();
 
 /**
  * Compile `node`.
@@ -69,24 +24,6 @@ Compiler.prototype.compile = function(node){
   return node.stylesheet
     .rules.map(this.visit, this)
     .join('');
-};
-
-/**
- * Visit `node`.
- */
-
-Compiler.prototype.visit = function(node){
-  return this[node.type](node);
-};
-
-Compiler.prototype.visitMap = function(nodes, join){
-  join = join || '';
-  var result = '';
-  for (var i = 0, length = nodes.length; i < length; i++) {
-    result += this.visit(nodes[i]);
-    if (i < length - 1) result += this.emit(join);
-  }
-  return result;
 };
 
 /**
@@ -112,7 +49,7 @@ Compiler.prototype.import = function(node){
 Compiler.prototype.media = function(node){
   return this.emit('@media ' + node.media, node.position, true)
     + this.emit('{')
-    + this.visitMap(node.rules)
+    + this.mapVisit(node.rules)
     + this.emit('}');
 };
 
@@ -125,7 +62,7 @@ Compiler.prototype.document = function(node){
 
   return this.emit(doc, node.position, true)
     + this.emit('{')
-    + this.visitMap(node.rules)
+    + this.mapVisit(node.rules)
     + this.emit('}');
 };
 
@@ -152,7 +89,7 @@ Compiler.prototype.namespace = function(node){
 Compiler.prototype.supports = function(node){
   return this.emit('@supports ' + node.supports, node.position, true)
     + this.emit('{')
-    + this.visitMap(node.rules)
+    + this.mapVisit(node.rules)
     + this.emit('}');
 };
 
@@ -166,7 +103,7 @@ Compiler.prototype.keyframes = function(node){
     + 'keyframes '
     + node.name, node.position, true)
     + this.emit('{')
-    + this.visitMap(node.keyframes)
+    + this.mapVisit(node.keyframes)
     + this.emit('}');
 };
 
@@ -179,7 +116,7 @@ Compiler.prototype.keyframe = function(node){
 
   return this.emit(node.values.join(','), node.position, true)
     + this.emit('{')
-    + this.visitMap(decls)
+    + this.mapVisit(decls)
     + this.emit('}');
 };
 
@@ -194,7 +131,7 @@ Compiler.prototype.page = function(node){
 
   return this.emit('@page ' + sel, node.position, true)
     + this.emit('{')
-    + this.visitMap(node.declarations)
+    + this.mapVisit(node.declarations)
     + this.emit('}');
 };
 
@@ -208,7 +145,7 @@ Compiler.prototype.rule = function(node){
 
   return this.emit(node.selectors.join(','), node.position, true)
     + this.emit('{')
-    + this.visitMap(decls)
+    + this.mapVisit(decls)
     + this.emit('}');
 };
