+++ lib/identity.js
@@ -1,4 +1,4 @@
-var SourceMapGenerator = require('source-map').SourceMapGenerator;
+var CompilerBase = require('./compiler');
 
 /**
  * Expose compiler.
@@ -12,55 +12,11 @@ module.exports = Compiler;
 
 function Compiler(options) {
   options = options || {};
-  this.map = new SourceMapGenerator({file: options.filename || 'generated.css'});
-  this.position = {line: 1, column: 0};
+  CompilerBase.call(this, options);
   this.indentation = options.indent;
 }
 
-Compiler.prototype.updatePosition = function(str) {
-  var lines = str.match(/\n/g);
-  if (lines) this.position.line += lines.length;
-  var i = str.lastIndexOf('\n');
-  this.position.column = ~i ? str.length - i : this.position.column + str.length;
-}
-
-/**
- * Emit string and update current position
- */
-
-Compiler.prototype.emit = function(str, pos, startOnly) {
-  if (pos && pos.start) {
-    this.map.addMapping({
-      generated: {
-        line: this.position.line,
-        column: Math.max(this.position.column - 1, 0)
-      },
-      source: 'source.css',
-      original: {
-        line: pos.start.line,
-        column: pos.start.column - 1
-      }
-    });
-  }
-
-  this.updatePosition(str);
-
-  if (!startOnly && pos && pos.end) {
-    this.map.addMapping({
-      generated: {
-        line: this.position.line,
-        column: Math.max(this.position.column - 1, 0)
-      },
-      source: 'source.css',
-      original: {
-        line: pos.end.line,
-        column: pos.end.column - 1
-      }
-    });
-  }
-
-  return str;
-}
+Compiler.prototype = new CompilerBase;
 
 /**
  * Compile `node`.
@@ -71,29 +27,11 @@ Compiler.prototype.compile = function(node){
 };
 
 /**
- * Visit `node`.
- */
-
-Compiler.prototype.visit = function(node){
-  return this[node.type](node);
-};
-
-Compiler.prototype.visitMap = function(nodes, join){
-  join = join || '';
-  var result = '';
-  for (var i = 0, length = nodes.length; i < length; i++) {
-    result += this.visit(nodes[i]);
-    if (i < length - 1) result += this.emit(join);
-  }
-  return result;
-};
-
-/**
  * Visit stylesheet node.
  */
 
 Compiler.prototype.stylesheet = function(node){
-  return this.visitMap(node.stylesheet.rules, '\n\n');
+  return this.mapVisit(node.stylesheet.rules, '\n\n');
 };
 
 /**
@@ -121,7 +59,7 @@ Compiler.prototype.media = function(node){
     + this.emit(
         ' {\n'
         + this.indent(1))
-    + this.visitMap(node.rules, '\n\n')
+    + this.mapVisit(node.rules, '\n\n')
     + this.emit(
         this.indent(-1)
         + '\n}');
@@ -139,7 +77,7 @@ Compiler.prototype.document = function(node){
         ' '
       + ' {\n'
       + this.indent(1))
-    + this.visitMap(node.rules, '\n\n')
+    + this.mapVisit(node.rules, '\n\n')
     + this.emit(
         this.indent(-1)
         + '\n}');
@@ -170,7 +108,7 @@ Compiler.prototype.supports = function(node){
     + this.emit(
       ' {\n'
       + this.indent(1))
-    + this.visitMap(node.rules, '\n\n')
+    + this.mapVisit(node.rules, '\n\n')
     + this.emit(
         this.indent(-1)
         + '\n}');
@@ -185,7 +123,7 @@ Compiler.prototype.keyframes = function(node){
     + this.emit(
       ' {\n'
       + this.indent(1))
-    + this.visitMap(node.keyframes, '\n')
+    + this.mapVisit(node.keyframes, '\n')
     + this.emit(
         this.indent(-1)
         + '}');
@@ -203,7 +141,7 @@ Compiler.prototype.keyframe = function(node){
     + this.emit(
       ' {\n'
       + this.indent(1))
-    + this.visitMap(decls, '\n')
+    + this.mapVisit(decls, '\n')
     + this.emit(
       this.indent(-1)
       + '\n'
@@ -222,7 +160,7 @@ Compiler.prototype.page = function(node){
   return this.emit('@page ' + sel, node.position, true)
     + this.emit('{\n')
     + this.emit(this.indent(1))
-    + this.visitMap(node.declarations, '\n')
+    + this.mapVisit(node.declarations, '\n')
     + this.emit(this.indent(-1))
     + this.emit('\n}');
 };
@@ -239,7 +177,7 @@ Compiler.prototype.rule = function(node){
   return this.emit(node.selectors.map(function(s){ return indent + s }).join(',\n'), node.position, true)
     + this.emit(' {\n')
     + this.emit(this.indent(1))
-    + this.visitMap(decls, '\n')
+    + this.mapVisit(decls, '\n')
     + this.emit(this.indent(-1))
     + this.emit('\n' + this.indent() + '}');
 };
